Refactor(cli): Split the features and wizard modules
features.py (757 lines) and wizard.py (687) both sat past the
400-line reviewability limit the aislop code-quality gate applies.

Split both into packages at the same dotted paths. features becomes
models (the FeatureInfo type and the AVAILABLE_FEATURES registry, 45%
of the original file), queries (registry reads) and formatting
(terminal rendering). wizard follows the banner comments the file
already carried: models (the three config templates), prompts,
configuration (the ConfigurationWizard class) and api (the public
entry points).

run_wizard now resolves ConfigurationWizard from cli.wizard.api, so
the two tests that patch it target that module; patching the package
attribute would no longer reach the global, and the real wizard would
block on stdin. A __main__.py restores 'python -m cli.wizard', which a
package __init__ cannot provide.
