+++ tests/test_cli/test_wizard.py
@@ -408,7 +408,7 @@ class TestConfigurationWizard:
 class TestRunWizard:
     """Test run_wizard convenience function."""
 
-    @patch("cli.wizard.ConfigurationWizard")
+    @patch("cli.wizard.api.ConfigurationWizard")
     def test_run_wizard_calls_wizard(self, mock_wizard_class, tmp_path):
         """Test run_wizard instantiates and runs wizard."""
         mock_wizard = MagicMock()
@@ -421,7 +421,7 @@ class TestRunWizard:
         mock_wizard_class.assert_called_once()
         mock_wizard.run.assert_called_once_with(str(output_path))
 
-    @patch("cli.wizard.ConfigurationWizard")
+    @patch("cli.wizard.api.ConfigurationWizard")
     def test_run_wizard_without_path(self, mock_wizard_class):
         """Test run_wizard without explicit path."""
         mock_wizard = MagicMock()
